Operand rewrite - replace all counted loops
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/immuse-rewrite-branch@89547 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/tree-ssa.c
@@ -594,7 +594,7 @@ verify_ssa (bool check_modified_stmt)
 	  if (check_modified_stmt && stmt_modified_p (stmt))
 	    {
 	      error ("Stmt (0x%x) marked modified after optimization pass : ",
-		     (unsigned long)stmt);
+		     (unsigned int)(size_t)stmt);
 	      print_generic_stmt (stderr, stmt, TDF_VOPS);
 	      goto err;
 	    }
